refactor!: fix namespaces, mark LamdaActivator internal but expose ctor getters
